changed temp return to float so that we can return NaN when there's no thermocouple

+++ max6675.h
@@ -8,8 +8,8 @@ class MAX6675 {
  public:
   MAX6675(int8_t SCLK, int8_t CS, int8_t MISO);
 
-    float readCelsius(void);
-    float readFarenheit(void);
+  double readCelsius(void);
+  double readFarenheit(void);
  private:
   int8_t sclk, miso, cs;
   uint8_t spiread(void);
